Fixed issues with moving auth_tests to directory
