+++ volttrontesting/platform/auth_tests/conftest.py
@@ -1,0 +1,7 @@
+def assert_auth_entries_same(e1, e2):
+    for field in ['domain', 'address', 'user_id', 'credentials', 'comments',
+                  'enabled']:
+        assert e1[field] == e2[field]
+    for field in ['roles', 'groups']:
+        assert set(e1[field]) == set(e2[field])
+    assert e1['capabilities'] == e2['capabilities']
